+++ src/app/(main)/page.tsx
@@ -1,7 +1,7 @@
 import React, { Suspense } from "react";
 
 import Loading from "@/app/(main)/loading";
-import HomeContent from "@/component/home";
+import HomeContent from "@/components/home";
 
 export default async function Home() {
 	return (
